[TEST] pyopenms test
